This seems as a very large commit. I will explain why this seems so.
First of all I prepared the project for the use of wicket. This
entails various dependencies in the project pom, addition of 
various directories and files.
Furthermore I do not like the default behavior of wicket to
search the corresponding html file in the same package
structure as the java file. So I created an
AnnotatedLocator.java which makes use of annotations to
direct the wicket container to the correct container.
Lastly I greeting a greeting file with a backing
implementation to test the working of setup.